fixed gcc warnings

+++ include/trurl/niobuf.h
@@ -19,7 +19,7 @@ int n_iobuf_write(tn_iobuf *iobuf, const void *buf, size_t size);
 int n_iobuf_flush(tn_iobuf *iobuf);
 int n_iobuf_close(tn_iobuf *iobuf);
 
-long n_iobuf_seek(tn_iobuf *iobuf, long offset, int whence);
+int n_iobuf_seek(tn_iobuf *iobuf, long offset, int whence);
 long n_iobuf_tell(tn_iobuf *iobuf);
 
 char *n_iobuf_gets(tn_iobuf *iobuf, char *dest, size_t size);
